Closely matching SQL DiscontItems-ContItems Query
Matching SQL Discontinuing items Query to SQL Continuing Items Query as close as possible. Although the the filters in the beginning worked correctly for the given continuing/discontinuing items, the conditional SQL statements should be as identical as possible for when continuing items is changed to a discontinuing status so that MS Access continue to query data correctly.